hometask-3 problems fixed and rollback added

+++ migration/src/main/resources/db/changelog/v.1.0.0/create-table-order-items.sql
@@ -5,7 +5,7 @@ create table if not exists order_items
     id bigint not null default nextval('order_items_seq'),
     order_id bigint not null,
     restaurant_menu_item bigint not null,
-    price int not null,
+    price numeric not null,
     quantity int not null,
     constraint order_items_pk primary key (id),
     constraint order_fk foreign key(order_id) references orders(id)
